changed AttachExchange variants
