+++ fe2o3-amqp/src/link/receiver_link.rs
@@ -1,7 +1,7 @@
 use fe2o3_amqp_types::messaging::message::DecodeIntoMessage;
 use serde_amqp::format_code::EncodingCodes;
 
-use crate::{util::{AsByteIterator, IntoReader}, AttachExchange};
+use crate::{util::{AsByteIterator, IntoReader}};
 
 use super::*;
